Start to implement class diagram
ref: https://github.com/LibreSign/libresign/pull/1754#issuecomment-1581109045
