+++ lib/Handler/CertificateEngine/AEngineHandler.php
@@ -1,0 +1,116 @@
+<?php
+
+namespace OCA\Libresign\Handler\CertificateEngine;
+
+use OCA\Libresign\AppInfo\Application;
+use OCA\Libresign\Exception\LibresignException;
+use OCA\Libresign\Helper\MagicGetterSetterTrait;
+use OCP\IConfig;
+
+/**
+ * @method IEngineHandler setPassword(string $password)
+ * @method string getPassword()
+ * @method IEngineHandler setCommonName(string $commonName)
+ * @method string getCommonName()
+ * @method IEngineHandler setHosts(array $hosts)
+ * @method array getHosts()
+ * @method IEngineHandler setFriendlyName(string $friendlyName)
+ * @method string getFriendlyName()
+ * @method IEngineHandler setCountry(string $country)
+ * @method string getCountry()
+ * @method IEngineHandler setState(string $state)
+ * @method string getState()
+ * @method IEngineHandler setLocality(string $locality)
+ * @method string getLocality()
+ * @method IEngineHandler setOrganization(string $organization)
+ * @method string getOrganization()
+ * @method IEngineHandler setOrganizationUnit(string $organizationUnit)
+ * @method string getOrganizationUnit()
+ * @method string getConfigPath()
+ */
+class AEngineHandler {
+	use MagicGetterSetterTrait;
+
+	protected string $commonName = '';
+	protected array $hosts = [];
+	protected string $friendlyName = '';
+	protected string $country = '';
+	protected string $state = '';
+	protected string $locality = '';
+	protected string $organization = '';
+	protected string $organizationUnit = '';
+	protected string $password = '';
+	protected string $configPath = '';
+	protected string $engine = '';
+
+	public function __construct(
+		protected IConfig $config
+	) {
+	}
+
+	public function generateCertificate(): string {
+		$certKeys = $this->newCert();
+		$certContent = null;
+		try {
+			openssl_pkcs12_export(
+				$certKeys['certificate'],
+				$certContent,
+				$certKeys['private_key'],
+				$this->getPassword(),
+				['friendly_name' => $this->getFriendlyName()],
+			);
+		} catch (\Throwable $th) {
+			throw new LibresignException('Error while creating certificate file', 500);
+		}
+
+		return $certContent;
+	}
+
+	public function translateToLong($name): string {
+		switch ($name) {
+			case 'CN':
+				return 'CommonName';
+			case 'C':
+				return 'Country';
+			case 'ST':
+				return 'State';
+			case 'L':
+				return 'Locality';
+			case 'O':
+				return 'Organization';
+			case 'OU':
+				return 'OrganizationUnit';
+		}
+		return '';
+	}
+
+	protected function setEngine(string $engine): void {
+		$this->config->setAppValue(Application::APP_ID, 'certificate_engine', $engine);
+		$this->engine = $engine;
+	}
+
+	protected function populateInstance(IEngineHandler $engine): IEngineHandler {
+		$rootCert = $this->config->getAppValue(Application::APP_ID, 'rootCert');
+		$rootCert = json_decode($rootCert, true);
+		if (!$rootCert) {
+			throw new LibresignException('Invalid or empty root certificate', 500);
+		}
+		if (!empty($rootCert['names'])) {
+			foreach ($rootCert['names'] as $id => $customName) {
+				$longCustomName = $this->translateToLong($id);
+				$engine->{'set' . ucfirst($longCustomName)}($customName['value']);
+			}
+		}
+		if (!$engine->getCommonName()) {
+			$engine->setCommonName($rootCert['commonName']);
+		}
+		if (!$engine->getConfigPath()) {
+			$engine->setConfigPath($this->config->getAppValue(Application::APP_ID, 'configPath'));
+		}
+		return $engine;
+	}
+
+	public function setConfigPath(string $configPath): void {
+		$this->config->setAppValue(Application::APP_ID, 'config_path', $configPath);
+	}
+}
